feat: added licenses
