+++ datasets_tagged/README.md
@@ -1,0 +1,82 @@
+```markdown
+# Tagged Datasets
+
+This folder contains the tagged datasets generated by the framework. Each file represents a dataset with metadata and tags that describe the quality and characteristics of the data.
+
+## File Format
+
+The tagged datasets are stored in JSON format. Each JSON file contains the following structure:
+
+### JSON Structure
+
+- **unitActions**: A list of actions performed by units in the dataset.
+- **gameStates**: A list of game states corresponding to the actions.
+- **gameBoard**: The representation of the game board.
+- **tags**: Tags generated for the dataset, providing additional metadata.
+- **notes**: Notes about the dataset, including the original file name.
+- **properties**: Metadata about the dataset:
+  - **id**: Unique identifier for the dataset.
+  - **quality**: Quality score of the dataset (0-100).
+  - **players**: Total number of players in the dataset.
+  - **bots**: Number of bot players in the dataset.
+  - **human_players**: Number of human players in the dataset.
+  - **board_size**: Size of the game board.
+  - **units**: Total number of units in the dataset.
+  - **bv**: Battle value of the dataset.
+
+### Example JSON
+
+```json
+{
+  "unitActions": [...],
+  "gameStates": [...],
+  "gameBoard": {...},
+  "tags": [...],
+  "notes": "original file name: example.tsv",
+  "properties": {
+    "id": 1,
+    "quality": 25,
+    "players": 4,
+    "bots": 2,
+    "human_players": 2,
+    "board_size": [32, 32],
+    "units": 10,
+    "bv": 8600
+  }
+}
+```
+
+## File Naming Convention
+
+The filenames of the tagged datasets follow a specific format to encode key characteristics of the dataset:
+
+```
+tagged dataset q=<quality> a=<actions> p=<players> u=<units> bv=<battle_value> d=<date> t=<time> id=<id>.json
+```
+
+### Filename Components
+
+- **q**: Quality score of the dataset (0-100).
+- **a**: Number of actions in the dataset.
+- **p**: Number of players in the dataset.
+- **u**: Number of units in the dataset.
+- **bv**: Total battle value of the dataset.
+- **d**: Date when the dataset was created (format: YYYY-MM-DD).
+- **t**: Local time when the dataset was created (format: HH-MM-SS).
+- **id**: Unique identifier for the dataset during tagging.
+
+### Example Filename
+
+```
+tagged dataset q=025 a=1200 p=4 u=10 bv=008600 d=2025-01-01 t=12-00-00 id=1.json
+```
+
+- **q=025**: Quality score is 25.
+- **a=1200**: The dataset contains 1200 actions.
+- **p=4**: There are 4 players in the dataset.
+- **u=10**: The dataset includes 10 units.
+- **bv=008600**: The total battle value is 8600.
+- **d=2025-01-01**: The dataset was created on January 1, 2025.
+- **t=12-00-00**: The dataset was created at 12:00:00.
+- **id=1**: The unique identifier for this dataset is 1.
+```
